+++ pkg/iam/iam.go
@@ -38,6 +38,7 @@ type IAMServiceConfig struct {
 	RoleType         string
 	Region           string
 	PrincipalRoleARN string
+	CustomTags       map[string]string
 
 	IAMClientFactory func(awsclientgo.ConfigProvider) iamiface.IAMAPI
 }
@@ -51,6 +52,7 @@ type IAMService struct {
 	region           string
 	roleType         string
 	principalRoleARN string
+	customTags       map[string]string
 }
 
 type Route53RoleParams struct {
@@ -91,6 +93,7 @@ func New(config IAMServiceConfig) (*IAMService, error) {
 		roleType:         config.RoleType,
 		region:           config.Region,
 		principalRoleARN: config.PrincipalRoleARN,
+		customTags:       config.CustomTags,
 	}
 
 	return s, nil
@@ -259,6 +262,12 @@ func (s *IAMService) createRole(roleName string, roleType string, params interfa
 			Key:   aws.String(fmt.Sprintf(ClusterIDTag, s.clusterName)),
 			Value: aws.String("owned"),
 		},
+	}
+	for k, v := range s.customTags {
+		tags = append(tags, &awsiam.Tag{
+			Key:   aws.String(k),
+			Value: aws.String(v),
+		})
 	}
 
 	_, err = s.iamClient.CreateRole(&awsiam.CreateRoleInput{